refactor: 스타일 컴포넌트인 NavRow 경로 수정
