+++ front/components/recycle/layout/PageLayout.tsx
@@ -1,15 +1,13 @@
 import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import styled from 'styled-components';
 import * as t from '../../../reducers/type';
 
-import AppLayout from '../../AppLayout';
 import Nav from '../../Nav';
 import Intersection from '../element/Intersection';
 
 import { media } from '../../../styles/media';
-
-import { NavRow } from '../../../pages/closet/overview';
+import { NavRow } from '../../main/state/RenderPageInLoading';
 
 type Props = {
   children: React.ReactNode;
