+++ front/components/state/error/RenderErrorPage.tsx
@@ -6,7 +6,7 @@ import { media } from '../../../styles/media';
 import AButton from '../../recycle/element/button/AButton';
 import Nav from '../../Nav';
 import Intersection from '../../recycle/element/Intersection';
-import { NavRow } from '../../../pages/closet/overview';
+import { NavRow } from '../../main/state/RenderPageInLoading';
 import { useLottiePropsByError, PropsByError } from './Data';
 
 const RenderErrorPage = ({ state }: PropsByError) => {
